+++ CHANGES.md
@@ -5,6 +5,7 @@ Changes in 8.5
 --------------
 - Fix SQL error duplicate key value violates unique constraint "food_service_menus_title" in Menus.php
 - SQL add PRIMARY KEY to STAFF_EXCEPTIONS table in rosariosis.sql
+- SQL PROFILE_EXCEPTIONS & STAFF_EXCEPTIONS tables: Add Admin Student Payments Delete restriction in Update.fnc.php & rosariosis.sql
 
 Changes in 8.4
 --------------